move deribit option_chain logic into provider class
