+++ backend/data/providers/__init__.py
@@ -1,0 +1,10 @@
+# This file marks the data directory as a Python package.
+
+# Importing relevant classes and functions
+from .DeribitProvider import DeribitProvider, Provider
+
+
+__all__ = [
+    "DeribitProvider",
+    "Provider",
+]
